@@ -32,27 +32,27 @@
   },
   "homepage": "https://github.com/ogix/ng2-fancy-image-uploader#readme",
   "peerDependencies": {
-    "@angular/common": "^4.0.0",
-    "@angular/core": "^4.0.0",
-    "@angular/forms": "^4.0.0",
-    "@angular/platform-browser": "^4.0.0"
+    "@angular/common": "^5.0.0",
+    "@angular/core": "^5.0.0",
+    "@angular/forms": "^5.0.0",
+    "@angular/platform-browser": "^5.0.0"
   },
   "dependencies": {
-    "@angular/common": "^4.0.0",
-    "@angular/compiler": "^4.0.0",
-    "@angular/core": "^4.0.0",
-    "@angular/forms": "^4.0.0",
-    "@angular/platform-browser": "^4.0.0",
-    "@types/cropperjs": "0.0.23",
+    "@angular/common": "^5.0.0",
+    "@angular/compiler": "^5.0.0",
+    "@angular/core": "^5.0.0",
+    "@angular/forms": "^5.0.0",
+    "@angular/platform-browser": "^5.0.0",
+    "@types/cropperjs": "1.1.0",
     "core-js": "^2.4.1",
-    "cropperjs": "^1.0.0-beta.1",
-    "rxjs": "^5.1.0",
+    "cropperjs": "^1.1.3",
+    "rxjs": "^5.5.2",
     "zone.js": "^0.8.4"
   },
   "devDependencies": {
-    "@angular/compiler-cli": "^4.0.0",
+    "@angular/compiler-cli": "^5.0.0",
     "@types/node": "~6.0.60",
-    "typescript": "~2.2.0"
+    "typescript": "~2.4.0"
   },
   "files": [
     "dist"
